Support large task definitions
When fetching a large Task definition with one of the resolvers, the
content gets truncated. This commit fixes the issue by updating the
underlying tektoncd/pipeline library to the latest version where the
actual fix was done.

As part of this commit, the "new" resolver calls are being used since
the previous ones are now marked as deprecated.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>